Added chore flow with owning user relation

+++ MyChores.Application/Features/Chores/Commands/DeleteChoreCommand .cs
@@ -10,25 +10,31 @@ using System.Threading.Tasks;
 
 namespace MyChores.Application.Features.Chores.Commands
 {
-    public class DeleteChoreCommand : IRequest
+
+    public class DeleteChoreForUserCommand : IRequest
     {
         public Guid Id { get; set; }
+        public string UserId { get; set; }
     }
 
-    public class DeleteChoreCommandHandler : IRequestHandler<DeleteChoreCommand>
+
+    public class DeleteChoreCommandHandler : IRequestHandler<DeleteChoreForUserCommand>
     {
         private readonly IMyChoresDbContext _context;
         public DeleteChoreCommandHandler(IMyChoresDbContext context) 
         {
             _context = context;
         }
-        public async Task Handle(DeleteChoreCommand command, CancellationToken cancellationToken)
+        public async Task Handle(DeleteChoreForUserCommand command, CancellationToken cancellationToken)
         {
 
             var entity = await _context.Chores.FindAsync(command.Id, cancellationToken);
 
             if (entity == null)
                 throw new Exception($"Can't find entity with id:{command.Id} to delete");
+
+            if(!entity.UserId.Equals(command.UserId))
+                throw new Exception($"Can't find entity with id:{command.Id} of user with id:{command.UserId} to delete");
 
             _context.Chores.Remove(entity);
